[rom_ctrl,rtl] Avoid some backwards part-selects
A part-select (defined in section 11.5.1 of IEEE 1800-2012) can either
be indexed on non-indexed. The code that was here before used a
non-indexed part-selects, which are defined as having the syntax
"vect[msb_expr:lsb_expr]". Rather confusingly, "msb" here means "the
left-most bit in the vector (as defined by it's data type). (see
section 6.9.1 of the same reference).

As such, the msb of a variable defined as "bit [0:10] foo" is the one
at index 0, not 10. Accessing foo[3:0] will seemingly work, but the
EDA tool might throw a warning about the fact that 3 can't be an MSB
if 0 is the LSB.

This is all rather mysterious! But we can cheat and make things easy
for ourselves by using an indexed part-select instead. The [0 +: AW]
range is equivalent to what we had before, but doesn't really care
about the order of the LSB/MSB in the type of the integral variables
that are being sliced.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>